db_writer/requirements.txt
kafka-python
cassandra-driver
pymemcache